Fix implicit 64-to-32 bit integer narrowing
Summary:
These fire when a 64-bit integer (e.g., `size_t`, `int64_t`) is implicitly converted to a 32-bit type (e.g., `int`, `uint32_t`), which can cause data loss.

Fixes applied:
- Widened variable types to match source (e.g., `int` → `size_t`)
- Added explicit `static_cast` where narrowing is intentional and safe

Differential Revision: D97079855

+++ momentum/solver/solver.cpp
@@ -43,7 +43,7 @@ void SolverT<T>::setEnabledParameters(const ParameterSet& parameterSet) {
   newParameterPattern_ = true;
 
   solverFunction_->setEnabledParameters(parameterSet);
-  actualParameters_ = solverFunction_->getActualParameters();
+  actualParameters_ = gsl::narrow<int>(solverFunction_->getActualParameters());
 }
 
 template <typename T>
